+++ Project.toml
@@ -12,19 +12,19 @@ FunSQL = "cf6cc811-59f4-4a10-b258-a8547a8f6407"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 LibPQ = "194296ae-ab2e-5f79-8cd4-7183a0a5a0d1"
 MLStyle = "d8e11817-5142-5d16-987a-aa16d5891078"
-MySQL = "39abe10b-433b-5dbd-92d4-e302a9df00cd"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SQLite = "0aa819cd-b072-5ff4-a722-6bc24af294d9"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
+MySQL = "39abe10b-433b-5dbd-92d4-e302a9df00cd"
 
 [weakdeps]
 DuckDB = "d2f5444f-75bc-4fdf-ac35-56f514c445e1"
 LibPQ = "2196b101-3ac0-4e12-b5a8-532424305d17"
 
 [extensions]
-DuckDBAlgRelExt = "DuckDB"
-LibPQAlgRelExt = "LibPQ"
-MySQLACSetsExt = "MySQL"
+# DuckDBAlgRelExt = "DuckDB"
+# LibPQAlgRelExt = "LibPQ"
+MySQLAlgRelExt = "MySQL"
 SQLiteAlgRelExt = "SQLite"
 
 [compat]
@@ -32,8 +32,6 @@ ACSets = "0.2.21"
 Catlab = "0.16"
 DBInterface = "2.6.1"
 DataFrames = "1.0"
-MLStyle = "0.4.17"
-MySQL = "1.4.7"
 Reexport = "1.2.2"
 SQLite = "1.6.1"
 julia = "1.6"
