+++ test/SQLiteInterop.jl
@@ -11,12 +11,12 @@ using SQLite
   (val!Salary, Name)::AttrType
   (Employee, Manager, Income, Salary)::Ob
   name::Attr(Employee, Name)
-
+  #
   (man!employee, man!manager)::Hom(Manager, Employee)
-
+  #
   inc!employee::Hom(Income, Employee)
   inc!salary::Hom(Income, Salary)
-
+  #
   sal!salary::Attr(Salary, val!Salary)
 end
 
@@ -24,6 +24,8 @@ busSchema = SQLSchema(Business; types = Dict(:val!Salary => Float64, :Name => St
 
 db = SQLite.DB()
 splt_stmts = split(render_schema(busSchema), "\n")
+
+vas = VirtualACSet(db)
 
 @testset "Generate DB Schema" begin
   for stmt in splt_stmts
